pyproject.toml
[tool.poetry]
name = "ssb-project-cli"
version = "1.0.15"
description = "SSB Project CLI"
authors = ["Statistics Norway <stat-dev@ssb.no>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/statisticsnorway/ssb-project-cli"
repository = "https://github.com/statisticsnorway/ssb-project-cli"
documentation = "https://ssb-project-cli.readthedocs.io"
packages = [
    { include = "ssb_project_cli", from = "src" },
]
classifiers = [
    "Development Status :: 4 - Beta",
]

[tool.poetry.urls]
Changelog = "https://github.com/statisticsnorway/ssb-project-cli/releases"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
click = ">=8.0.1"
dparse = "^0.6.2"
typer = "^0.6.1"
GitPython = "^3.1.27"
PyGithub = "^1.55"
cruft = "^2.11.1"
rich = "^12.5.1"
urllib3 = "^1.26.12"
requests = "^2.28.1"
types-toml = "^0.10.8"
questionary = "^1.10.0"
certifi = "^2022.12.7"
psutil = "^5.9.4"
types-urllib3 = "^1.26.25.10"
types-requests = "^2.28.11.17"
types-psutil = "^5.9.5.12"
kvakk-git-tools = "^2.2.1"


[tool.poetry.dev-dependencies]
Pygments = ">=2.10.0"
black = ">=21.10b0"
coverage = {extras = ["toml"], version = ">=6.2"}
darglint = ">=1.8.1"
flake8 = ">=4.0.1"
flake8-bandit = ">=2.1.2"
flake8-bugbear = ">=21.9.2"
flake8-docstrings = ">=1.6.0"
flake8-rst-docstrings = ">=0.2.5"
furo = ">=2021.11.12"
isort = ">=5.10.1"
mypy = ">=0.930"
pep8-naming = ">=0.12.1"
pre-commit = ">=2.16.0"
pre-commit-hooks = ">=4.1.0"
pytest = ">=7.2.0"
pyupgrade = ">=2.29.1"
safety = ">=1.10.3"
sphinx = ">=4.3.2"
sphinx-autobuild = ">=2021.3.14"
sphinx-click = ">=3.0.2"
typeguard = ">=2.13.3"
xdoctest = {extras = ["colors"], version = ">=0.15.10"}
myst-parser = {version = ">=0.16.1"}
mock = "^4.0.3"
mkdocs = "^1.4.0"
mkdocstrings-python = "^0.7.1"
mkdocs-literate-nav = "^0.5.0"
mkdocs-section-index = "^0.3.4"
mkdocstrings = "^0.19.0"
mkdocs-material = "^8.5.6"
mkdocs-click = "^0.8.0"
mkdocs-gen-files = "^0.4.0"
build = "^0.8.0"
Jinja2 = "^3.1.2"


[tool.poetry.scripts]
ssb-project = "ssb_project_cli.__main__:main"

[tool.poetry.group.dev.dependencies]
nox = "^2022.11.21"
nox-poetry = "^1.0.2"

[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
branch = true
source = ["ssb_project_cli", "tests"]

[tool.coverage.report]
show_missing = true
fail_under = 80

[tool.isort]
profile = "black"
force_single_line = true
lines_after_imports = 2

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
